requirements.txt
mcp[cli]==1.10.1
openai==1.75.0
python-dotenv
ipykernel
httpx
anthropic